Criação de mais botoes no menu de cadastro e modificacao nos packages para abordar a visualização de maneira separada
