+++ src/Main.java
@@ -1,4 +1,6 @@
-public class Main extends MainController{
+import View.View;
+
+public class Main extends View {
     public static void main(String[] args) {
         chamaMenuPrincipal();
     }
